Split desktop state and test support modules
