+++ crates/defra-agent-desktop/src/app/tests/live/mod.rs
@@ -8,7 +8,6 @@ pub(crate) use operator_config::*;
 mod chat;
 mod chat_followup;
 mod logs;
-mod operator_roundtrip;
 mod operator_scheduled;
 mod operator_switching;
 mod replication;
