+++ crates/defra-agent-desktop/src/app/tests/support/live_fixture/mod.rs
@@ -1,0 +1,14 @@
+use super::*;
+
+mod bootstrap;
+mod builders;
+mod types;
+
+pub(crate) use builders::{
+    build_live_desktop_fixture, build_multi_agent_desktop_fixture_with_backend,
+    build_multi_agent_live_desktop_fixture,
+};
+pub(crate) use types::{
+    live_deployment_case, LiveAgentDocs, LiveDeploymentCase, LiveDesktopFixture,
+    LiveRemoteDeployment, LiveSubmissionCase, MultiAgentLiveDesktopFixture,
+};
