+++ crates/defra-agent-desktop/src/app/tests/support/seed/mod.rs
@@ -1,0 +1,11 @@
+use super::*;
+
+mod operator;
+mod query;
+mod transcript;
+
+pub(crate) use operator::{
+    insert_agent_principal, seed_live_operator_documents, seed_operator_documents,
+};
+pub(crate) use query::query_has_row_by_unique_field;
+pub(crate) use transcript::insert_chat_transcript_documents;
