+++ crates/defra-agent-desktop/src/state/mod.rs
@@ -1,0 +1,16 @@
+mod activity;
+mod chat;
+mod operator;
+mod peers;
+mod shell;
+mod status;
+
+pub use self::activity::{Activity, PendingChatAction, PendingShellAction};
+pub use self::chat::{ChatEditorState, ChatShellState, ChatState, ToolDetailModalState};
+pub use self::operator::{
+    BackendDraft, BehaviorDraft, InferenceProfileDraft, OperatorDraft, OperatorSection,
+    OperatorState, ScheduledTaskDraft, ToolSelectionDraft,
+};
+pub use self::peers::PeersState;
+pub use self::shell::ShellState;
+pub use self::status::{IdentityState, LogsFilter, LogsState, OnboardingState, StatusBarState};
